% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.xlsf_data}
\alias{print.xlsf_data}
\title{print methods for xlsf object
special print for just data
print for just survey data
print for full objefct including data and questionnnaire}
\usage{
\method{print}{xlsf_data}(df)
}
\description{
print methods for xlsf object
special print for just data
print for just survey data
print for full objefct including data and questionnnaire
}
