% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_dict.R
\name{xlsf_dict}
\alias{xlsf_dict}
\title{xlsf_dict}
\usage{
xlsf_dict(xlsf)
}
\arguments{
\item{xlsf}{XLSForm entered as list of 2 named data frame (`survey`,`choices`)}
}
\description{
create nested look up table dictionary for xlsForm
}
\details{
need the list to be named
need label column to be chosen in teh creation of the xlsf object
}
\author{
Zack Arno
}
