% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_factorize.R
\name{xlsf_factorize}
\alias{xlsf_factorize}
\title{xlsf factorize - maybe call xlsf_relevel or xlsf_expand
factorize xlsf according to tool. Convert all categoric variables to factors and re-level so all levels are present.}
\usage{
xlsf_factorize(xlsf)
}
\description{
xlsf factorize - maybe call xlsf_relevel or xlsf_expand
factorize xlsf according to tool. Convert all categoric variables to factors and re-level so all levels are present.
}
