% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_relevel.R
\name{xlsf_fct_so}
\alias{xlsf_fct_so}
\title{xlsf_fct_so}
\usage{
xlsf_fct_so(df, q_name, xlsf)
}
\arguments{
\item{df}{XLSForm data set (i.e kobo/ODK)}

\item{q_name}{question to be refactored}

\item{xlsf}{object of class xlsf}
}
\description{
relevel select ones to questionnaire options. Designed for use in dplyr::across
}
\author{
Zack Arno
}
