% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_relevel.R
\name{xlsf_get_sm}
\alias{xlsf_get_sm}
\title{xlsf_get_sm}
\usage{
xlsf_get_sm(xlsf = kobo, sm_sep = "/")
}
\arguments{
\item{xlsf}{object of class= xlsf}

\item{sm_sep}{character used to sepearate select multiple question/choices (default=/)}
}
\description{
retrieve column names of all select one questions in questionnaire
}
\author{
Zack Arno
}
