% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_load.R
\name{xlsf_load}
\alias{xlsf_load}
\title{xlsf_load}
\usage{
xlsf_load(survey, choices, label = "label::English (en)")
}
\arguments{
\item{survey}{data.frame containing survey tab contents of xlsform}

\item{choices}{data.frame containing choices tab contents of xlsform}

\item{label}{colname that contians labels (default=label::English (en))}

\item{data}{list of data sets derived from form (main data set and any repeat group sets)
`TODO` need to incorporate data arguments -- maybe ellipses for more than one argument}
}
\description{
create xlsf object
}
\author{
Zack Arno
}
