% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{xlsf_mutate_label}
\alias{xlsf_mutate_label}
\title{mutate labels based on 1 column containing question name and other containing question value
typically used to label long format analyzed data}
\usage{
xlsf_mutate_label(df, q_names, q_values)
}
\description{
mutate labels based on 1 column containing question name and other containing question value
typically used to label long format analyzed data
}
