% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_relevel.R
\name{xlsf_relevel}
\alias{xlsf_relevel}
\title{xlsf_relevel}
\usage{
xlsf_relevel(df, xlsf, sm_sep = "/")
}
\arguments{
\item{df}{XLSForm data set (i.e kobo/ODK)}

\item{xlsf}{xlsf class object}
}
\description{
relevel data to XLSForm. fct_expand applied to relevel all select questions to choices in questionnaire. all select_multiple converted to logical
}
\author{
Zack Arno
}
