% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsf_query.R
\name{xlsf_query}
\alias{xlsf_query}
\title{xlsf_query}
\usage{
xlsf_query(xlsf, lab_query)
}
\arguments{
\item{xlsf}{XLSForm entered as list of 2 named data frame (`survey`,`choices`)}

\item{lab_query}{question label queryy}
}
\description{
quickly query throu kobo forms. Goal: minmize time you have to spend opening an checking XLSForm while processingn data in R
}
\author{
Zack Arno
}
